refactored code

+++ src/main/java/com/bondarenko/movieland/repository/GenreRepository.java
@@ -4,8 +4,6 @@ import com.bondarenko.movieland.entity.Genre;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
-import java.util.List;
-
 @Repository
-public interface GenreRepository extends JpaRepository<Genre, Integer> {
+public interface GenreRepository extends JpaRepository<Genre, Long> {
 }
